Project_1/requirements.txt
flask==3.0.2
flask-cors==4.0.1
langchain-core==0.2.5
langchain-google-genai==1.0.3
google-generativeai==0.5.3
python-dotenv==1.0.1
vercel-python-wsgi==0.2.5